LayoutTests/inspector/debugger/debugger-set-breakpoint-regex.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    foo();
}

function foo()
{
}

function test()
{
    InspectorTest.runDebuggerTestSuite([
        function testSetNoneOfURLAndRegex(next)
        {
            var url = "debugger-set-breakpoint.js";
            var urlRegex = "debugger-set-breakpoint.*";
            DebuggerAgent.setBreakpointByUrl(undefined, undefined, 1, step2);

            function step2(result)
            {
                InspectorTest.dump(result);
                next();
            }
        },

        function testSetBothURLAndRegex(next)
        {
            var url = "debugger-set-breakpoint.js";
            var urlRegex = "debugger-set-breakpoint.*";
            DebuggerAgent.setBreakpointByUrl(url, urlRegex, 1, step2);

            function step2(result)
            {
                InspectorTest.dump(result);
                next();
            }
        },

        function testSetByRegex(next)
        {
            var url = undefined;
            var urlRegex = "debugger-set-breakpoint.*";
            var columnNumber = undefined;
            var condition = undefined;
            DebuggerAgent.setBreakpointByUrl(url, urlRegex, 8, columnNumber, condition, step2);

            function step2(result)
            {
                InspectorTest.runTestFunctionAndWaitUntilPaused(step3);
            }
        
            function step3(callFrames)
            {
                InspectorTest.captureStackTrace(callFrames);
                next();
            }
        }
    ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests Debugger.setBreakpointByUrl with isRegex set to true.
</p>
</body>
</html>
