LayoutTests/inspector/debugger/debugger-set-breakpoint-regex-expected.txt

Tests Debugger.setBreakpointByUrl with isRegex set to true.

Debugger was enabled.

Running: testSetNoneOfURLAndRegex
{
    code : -32000
    message : "Either url or urlRegex must be specified."
}

Running: testSetBothURLAndRegex
{
    code : -32000
    message : "Either url or urlRegex must be specified."
}

Running: testSetByRegex
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (debugger-set-breakpoint-regex.html:9)
Script execution resumed.
Debugger was disabled.
